s1543: wait for queued/running flush threads to finish
The worry is that unmounting the filesystem may not properly
flush nodes that are sitting in the queue. This was probably not
a problem before since FUSE seems to detect running flushes
and doesn't allow you to unmount. However, it won't know
about queued flushes. The FUSE/vofs destroy method should
now block until the queue is empty.