t57858: fix handling of certfile
mountvofs passes the certfile from the command-line to the
Connection constructor. Later, the Connection object is
passed to the Client constructor.

Other scripts (e.g., vcp), instead call the Client constructor
without a Connection, and instead supply the certfile there.

The mountvofs case was not being handled correctly, and values
of certfile specified on the command-line were getting
overriden by the default value in the Client constructor
(~/.ssl/cadcproxy.pem)

In this commit the logic to the Client constructor is modified
so that certfile is taken from the connection, unless certfile
is specified explicitly as a parameter (in which case it
overrides the supplied connection object)