s1689: add cookie identification method
Mostly plumbing to pass the new "cookie" argument down to the
Client objects. If present, the string is currently just
added as a CADC_SSO cookie to request headers to mimic the single
sign-on behaviour with browsers. For now the presence of a
cookie will override a supplied cert, since http will be used
(which is faster than https required for certs).