t64562: add test to demonstrate Content-Length bug large files
Apparently there is a Java bug on the server side that limits
Content-Length to a signed 32-bit integer (about 2 gig files).
The only place this seems to cause a problem is when reading
a file larger than this limit using vofs, in which case the file
length is reported as 0, causing an exception.