Make md5==None not be a condition that aborts the read of a file.
The cache is built when a file in cache doesn't match the version on disk.  Previously the md5 on disk was being ignored and the
value set to None, thus every file read triggered a call to VOspace, which was a problem. That problem was avoided by putting md5==None
 as being a condition that DOES NOT load the from VOSpace, but that creates a race condition where the MD5sum is set to None in one thread
 while the other thread was attempting to build the metaData object.  Now we accept that MD5 sum on disk, this saves some vospace
 loops and avoids the race.  That MD5 sum must still match the VOSpace value and is just the value of the cached version of the file.