Refactored into some more modular code chunks.
Endpoints handles the customization of URLs to support multiple servers.
move method changed to a 'transfer' call with changes to transfer to enable that.
time conversion from VOSpace times to UNIX time placed in a static method.
Various PEP-8 cleanups.
Added locks to the  _get and _put methods so that they notify the suspend the cache.