Allow copy to follow the content-disposition setting from vospace instead of specified destination.
When a user sets destination=None or disposition = True then write the content to location specified in content-disposition.  Also return the string from content-disposition to the caller so the string can be used in various download mapping schemes.

This update also addresses a number of PEP-8 issues and maps some of the string types to try and be more straightforward (ie, understandable by PyCharm).