Undefined variable bug and pattern match fail
In the previous commit the chance that the variable 'cutout' might
be undefined, because this isn't a cutout call, was overlooked. Just
catch this possibility and ignore the cutout building step.

Also, the new path matching pattern is a bit more generic but also
weaker if the path contains characters that are not normal so the path
needed to be normalized first.